+++ Sources/CommonMark/Supporting Types/Children.swift
@@ -58,7 +58,13 @@ extension Document: ContainerOfBlocks {}
 extension BlockQuote: ContainerOfBlocks {}
 
 extension ContainerOfBlocks {
-    /// The block's children.
+    /**
+     The block elements contained by the node.
+
+     - Important: The returned child nodes are valid only during the lifetime of their parent.
+                  Use the `removeChildren()` method to detach and access children
+                  beyond the lifetime of their parent.
+     */
     public var children: [Block & Node] {
         get {
             return Children(of: self).compactMap { $0 as? Block & Node }
@@ -76,7 +82,7 @@ extension ContainerOfBlocks {
     }
 
     /**
-     Adds a block to the beginning of the block's children.
+     Adds a block to the beginning of the node's children.
 
      - Parameters:
         - child: The block to add.
@@ -88,7 +94,7 @@ extension ContainerOfBlocks {
     }
 
     /**
-     Adds a block to the end of the block's children.
+     Adds a block to the end of the node's children.
 
      - Parameters:
         - child: The block to add.
@@ -100,7 +106,7 @@ extension ContainerOfBlocks {
     }
 
     /**
-     Inserts a block to the block's children before a specified sibling.
+     Inserts a block to the node's children before a specified sibling.
 
      - Parameters:
         - child: The block to add.
@@ -113,7 +119,7 @@ extension ContainerOfBlocks {
     }
 
     /**
-     Inserts a block to the block's children after a specified sibling.
+     Inserts a block to the node's children after a specified sibling.
 
      - Parameters:
         - child: The block to add.
@@ -126,7 +132,7 @@ extension ContainerOfBlocks {
     }
 
     /**
-     Removes a block from the block's children.
+     Removes a block from the node's children.
 
      - Parameters:
         - child: The block to remove.
@@ -135,9 +141,26 @@ extension ContainerOfBlocks {
     @discardableResult
     public func remove(child: Block & Node) -> Bool {
         guard child.parent == self else { return false }
-        cmark_node_unlink(child.cmark_node)
-        child.managed = true
+        child.unlink()
+
         return true
+    }
+
+    /**
+     Removes and returns the node's children.
+
+     - Returns: An array of block structure elements.
+     */
+    @discardableResult
+    public func removeChildren() -> [Block & Node] {
+        var children: [Block & Node] = []
+
+        for child in self.children {
+            guard remove(child: child) else { continue }
+            children.append(child)
+        }
+
+        return children
     }
 }
 
@@ -155,7 +178,13 @@ extension Emphasis: ContainerOfInlineElements {}
 extension Link: ContainerOfInlineElements {}
 
 extension ContainerOfInlineElements {
-    /// The block's children.
+    /**
+     The inline elements contained by the node.
+
+     - Important: The returned child nodes are valid only during the lifetime of their parent.
+                  Use the `removeChildren()` method to detach and access children
+                  beyond the lifetime of their parent.
+     */
     public var children: [Inline & Node] {
         get {
             return Children(of: self).compactMap { $0 as? Inline & Node }
@@ -173,7 +202,7 @@ extension ContainerOfInlineElements {
     }
 
     /**
-     Adds an inline element to the beginning of the block's children.
+     Adds an inline element to the beginning of the node's children.
 
      - Parameters:
         - child: The inline element to add.
@@ -185,7 +214,7 @@ extension ContainerOfInlineElements {
     }
 
     /**
-     Adds an inline element to the end of the block's children.
+     Adds an inline element to the end of the node's children.
 
      - Parameters:
         - child: The inline element to add.
@@ -197,7 +226,7 @@ extension ContainerOfInlineElements {
     }
 
     /**
-     Inserts an inline element to the block's children before a specified sibling.
+     Inserts an inline element to the node's children before a specified sibling.
 
      - Parameters:
         - child: The inline element to add.
@@ -210,7 +239,7 @@ extension ContainerOfInlineElements {
     }
 
     /**
-     Inserts an inline element to the block's children after a specified sibling.
+     Inserts an inline element to the node's children after a specified sibling.
 
      - Parameters:
         - child: The inline element to add.
@@ -223,7 +252,7 @@ extension ContainerOfInlineElements {
     }
 
     /**
-     Removes an inline element from the block's children.
+     Removes an inline element from the node's children.
 
      - Parameters:
         - child: The inline element to remove.
@@ -232,16 +261,39 @@ extension ContainerOfInlineElements {
     @discardableResult
     public func remove(child: Inline & Node) -> Bool {
         guard child.parent == self else { return false }
-        cmark_node_unlink(child.cmark_node)
-        child.managed = true
+        child.unlink()
+
         return true
+    }
+
+    /**
+     Removes and returns the node's children.
+
+     - Returns: An array of inline content elements.
+     */
+    @discardableResult
+    public func removeChildren() -> [Inline & Node] {
+        var children: [Inline & Node] = []
+
+        for child in self.children {
+            guard remove(child: child) else { continue }
+            children.append(child)
+        }
+
+        return children
     }
 }
 
 // MARK: -
 
 extension List {
-    /// The block's children.
+    /**
+     The list's items.
+
+     - Important: The returned child nodes are valid only during the lifetime of their parent.
+                  Use the `removeChildren()` method to detach and access children
+                  beyond the lifetime of their parent.
+     */
     public var children: [Item] {
         get {
             return Children(of: self).compactMap { $0 as? Item }
@@ -259,10 +311,10 @@ extension List {
     }
 
     /**
-     Adds a block to the beginning of the block's children.
+     Adds an item to the beginning of the list.
 
      - Parameters:
-        - child: The block to add.
+        - child: The item to add.
      - Returns: `true` if successful, otherwise `false`.
      */
     @discardableResult
@@ -271,10 +323,10 @@ extension List {
     }
 
     /**
-     Adds a block to the end of the block's children.
+     Adds an to the end of the list.
 
      - Parameters:
-        - child: The block to add.
+        - child: The item to add.
      - Returns: `true` if successful, otherwise `false`.
     */
     @discardableResult
@@ -283,11 +335,11 @@ extension List {
     }
 
     /**
-     Inserts a block to the block's children before a specified sibling.
+     Inserts an item to the list before a specified sibling.
 
      - Parameters:
-        - child: The block to add.
-        - sibling: The child before which the block is added
+        - child: The item to add.
+        - sibling: The item before which the new item is added
      - Returns: `true` if successful, otherwise `false`.
     */
     @discardableResult
@@ -296,11 +348,11 @@ extension List {
     }
 
     /**
-     Inserts a block to the block's children after a specified sibling.
+     Inserts an item to the list after a specified sibling.
 
      - Parameters:
-        - child: The block to add.
-        - sibling: The child after which the block is added
+        - child: The item to add.
+        - sibling: The item after which the new item is added
      - Returns: `true` if successful, otherwise `false`.
     */
     @discardableResult
@@ -309,25 +361,48 @@ extension List {
     }
 
     /**
-     Removes a block from the block's children.
+     Removes an item from the list.
 
      - Parameters:
-        - child: The block to remove.
+        - child: The item to remove.
      - Returns: `true` if successful, otherwise `false`.
      */
     @discardableResult
     public func remove(child: Item) -> Bool {
-         guard child.parent == self else { return false }
-         cmark_node_unlink(child.cmark_node)
-         child.managed = true
-         return true
+        guard child.parent == self else { return false }
+        child.unlink()
+
+        return true
+    }
+
+    /**
+     Removes and returns the list's items.
+
+     - Returns: An array of list items.
+     */
+    @discardableResult
+    public func removeChildren() -> [Item] {
+        var children: [Item] = []
+
+        for child in self.children {
+            guard remove(child: child) else { continue }
+            children.append(child)
+        }
+
+        return children
     }
 }
 
 // MARK: -
 
 extension List.Item {
-    /// The list item's children.
+    /**
+     The elements contained by the list item.
+
+     - Important: The returned child nodes are valid only during the lifetime of their parent.
+                  Use the `removeChildren()` method to detach and access children
+                  beyond the lifetime of their parent.
+     */
     public var children: [Node] {
         get {
             return Children(of: self).map { $0 }
@@ -348,7 +423,7 @@ extension List.Item {
      Adds a node to the beginning of the list item's children.
 
      - Parameters:
-        - child: The block to add.
+        - child: The node to add.
      - Returns: `true` if successful, otherwise `false`.
      */
     @discardableResult
@@ -360,7 +435,7 @@ extension List.Item {
      Adds a node to the end of the list item's children.
 
      - Parameters:
-        - child: The block to add.
+        - child: The node to add.
      - Returns: `true` if successful, otherwise `false`.
     */
     @discardableResult
@@ -372,8 +447,8 @@ extension List.Item {
      Inserts a node to the list item's children before a specified sibling.
 
      - Parameters:
-        - child: The block to add.
-        - sibling: The child before which the block is added
+        - child: The node to add.
+        - sibling: The child before which the node is added
      - Returns: `true` if successful, otherwise `false`.
     */
     @discardableResult
@@ -385,8 +460,8 @@ extension List.Item {
      Inserts a node to the list item's children after a specified sibling.
 
      - Parameters:
-        - child: The block to add.
-        - sibling: The child after which the block is added
+        - child: The node to add.
+        - sibling: The child after which the node is added
      - Returns: `true` if successful, otherwise `false`.
     */
     @discardableResult
@@ -398,14 +473,30 @@ extension List.Item {
      Removes a node from the list item's children.
 
      - Parameters:
-        - child: The block to remove.
+        - child: The node to remove.
      - Returns: `true` if successful, otherwise `false`.
      */
     @discardableResult
     public func remove(child: Node) -> Bool {
         guard child.parent == self else { return false }
-        cmark_node_unlink(child.cmark_node)
-        child.managed = true
+        child.unlink()
+
         return true
+    }
+
+    /**
+     Removes and returns the list item's children.
+
+     - Returns: An array of nodes.
+     */
+    public func removeChildren() -> [Node] {
+        var children: [Node] = []
+
+        for child in self.children {
+            guard remove(child: child) else { continue }
+            children.append(child)
+        }
+
+        return children
     }
 }