+++ Tests/CommonMarkTests/DocumentCodingTests.swift
@@ -7,10 +7,10 @@ final class DocumentCodingTests: XCTestCase {
 
         // Workaround for "Top-level Document encoded as string JSON fragment."
         let encoded = try JSONEncoder().encode([document])
-        let decoded = try JSONDecoder().decode([Document].self, from: encoded)
+        let decoded = try JSONDecoder().decode([Document].self, from: encoded)[0]
 
-        let actual = decoded[0].render(format: .html)
-        let expected = document.render(format: .html)
+        let actual = decoded.render(format: .commonmark)
+        let expected = document.render(format: .commonmark)
 
         XCTAssertEqual(actual, expected)
     }
@@ -22,10 +22,10 @@ final class DocumentCodingTests: XCTestCase {
 
         // Workaround for "Top-level Heading encoded as string JSON fragment."
         let encoded = try JSONEncoder().encode([heading])
-        let decoded = try JSONDecoder().decode([Heading].self, from: encoded)
+        let decoded = try JSONDecoder().decode([Heading].self, from: encoded)[0]
 
-        let actual = try Document(decoded[0].description).render(format: .html)
-        let expected = try Document(heading.description).render(format: .html)
+        let actual = try Document(decoded.description).render(format: .commonmark)
+        let expected = try Document(heading.description).render(format: .commonmark)
 
         XCTAssertEqual(actual, expected)
     }
@@ -38,10 +38,10 @@ final class DocumentCodingTests: XCTestCase {
 
         // Workaround for "Top-level Link encoded as string JSON fragment."
         let encoded = try JSONEncoder().encode([link])
-        let decoded = try JSONDecoder().decode([Link].self, from: encoded)
+        let decoded = try JSONDecoder().decode([Link].self, from: encoded)[0]
 
-        let actual = try Document(decoded[0].description).render(format: .html)
-        let expected = try Document(link.description).render(format: .html)
+        let actual = try Document(decoded.description).render(format: .commonmark)
+        let expected = try Document(link.description).render(format: .commonmark)
 
         XCTAssertEqual(actual, expected)
     }